refactor(LottoController): Match 파일이 LottoController 역할을 수행하여 폴더 구조를 재정리하고 파일명을 변경하여 코드 가독성을 개선
